BK-#105- removing the ability to delete a requestoffer if it's been taken or closed.  Also, I noticed something - when I had declare exit handlers, it was making my user sqlexceptions disappear.  It gets pretty complicated if you want to send the handler your message, so maybe the point is, yeah, locks can occur if you write crappy code, but putting in exception handling code to handle your own errors should not be good practice.
Also, a minor note - I read in the MySQL documentation that specifically
the code 45000 is to be used for user sqlexceptions, nothing else.

+++ web_container/apache-tomcat-8.0.14/webapps/xenos/my_profile.jsp
@@ -26,11 +26,12 @@
 			<div class="requestoffer mine">
 				<a href="requestoffer.jsp?requestoffer=<%=r.requestoffer_id %>"> 
 					<%=Utils.safe_render(r.title)%> </a>
-				<a 
-					class="button" 
-					href="requestoffer.jsp?requestoffer=<%=r.requestoffer_id%>&amp;delete=true">
-					<%=loc.get(21,"Delete")%>
-				</a>
+        <%if(r.status == 76) {%>
+          <a class="button" 
+            href="requestoffer.jsp?requestoffer=<%=r.requestoffer_id%>&amp;delete=true">
+            <%=loc.get(21,"Delete")%>
+          </a>
+          <%}%>
 			</div>
 		<% } %>
 		</div>
